+++ changelog.md
@@ -1,9 +1,13 @@
 # Changelog
 
 ## [Unreleased]
+
 ### Added
 
 ### Changed
+
+- Update base Monarcula base colors
+- Update base Monarcula Soft base colors
 
 ### Deprecated
 
@@ -12,14 +16,21 @@
 ### Fixed
 
 ### Security
+
 ## [1.0.3]
+
 ### Fixed
+
 - Remove upper Intellij version limitation
 
 ## [1.0.2]
+
 ### Changed
+
 - Increase Intellij version to 2021.2
 
 ## [1.0.1]
+
 ### Added
+
 - Add color scheme support for golang